+++ src/github/index.ts
@@ -16,6 +16,7 @@ export * from "./workflow-jobs";
 export * from "./workflow-steps";
 
 export * from "./task-workflow";
+export * from "./task-workflow-job";
 
 export * from "./github-project";
